requirements.txt
Flask==2.0.1
tensorflow==2.12.0
gunicorn==20.1.0
opencv-python-headless==4.5.3.56
numpy>=1.19.2
keras==2.12.0
bcrypt==3.2.0
PyJWT==2.1.0
mysql-connector-python==8.0.26